+++ src/module/data/effect/base-effect.mjs
@@ -32,36 +32,12 @@ export default class BaseEffectModel extends foundry.data.ActiveEffectTypeDataMo
     const schema = super.defineSchema();
 
     const fields = foundry.data.fields;
-    const config = ds.CONFIG;
 
     schema.end = new fields.SchemaField({
-      type: new fields.StringField({ choices: config.effectEnds, blank: true, required: true }),
       roll: new FormulaField({ initial: "1d10 + @combat.save.bonus" }),
     });
 
     return schema;
-  }
-
-  /* -------------------------------------------------- */
-
-  /**
-   * An effect is also temporary if it has the `end` property set even though they have indeterminate lengths.
-   * @returns {boolean | null}
-   * @internal
-   */
-  get _isTemporary() {
-    if (this.end.type) return true;
-    else return null;
-  }
-
-  /* -------------------------------------------------- */
-
-  /**
-   * Returns the duration label appropriate to this model's `end` property.
-   * @returns {string}
-   */
-  get durationLabel() {
-    return ds.CONFIG.effectEnds[this.end.type]?.abbreviation ?? "";
   }
 
   /* -------------------------------------------------- */
@@ -74,25 +50,6 @@ export default class BaseEffectModel extends foundry.data.ActiveEffectTypeDataMo
     const target = this.parent.target ?? {};
     const invalidTypes = this.constructor.metadata[`invalid${target.documentName}Types`] ?? [];
     return invalidTypes.includes(target.type);
-  }
-
-  /* -------------------------------------------------- */
-
-  /** @import { ActiveEffectDuration, EffectDurationData } from "./_types" */
-
-  /**
-   * Subtype specific duration calculations.
-   * @returns {Omit<ActiveEffectDuration, keyof EffectDurationData> | null}
-   * @internal
-   */
-  get _prepareDuration() {
-    if (!this.end.type) return null;
-    return {
-      type: "draw-steel",
-      duration: null,
-      remaining: null,
-      label: this.durationLabel,
-    };
   }
 
   /* -------------------------------------------------- */
@@ -130,6 +87,7 @@ export default class BaseEffectModel extends foundry.data.ActiveEffectTypeDataMo
    * @param {object} [messageOptions={}]  Additional options which modify the created message.
    * @returns {Promise<DrawSteelChatMessage|object>} A promise which resolves to the created ChatMessage document if create is
    *                                                 true, or the Object of prepared chatData otherwise.
+   * TODO: Move to DrawSteelActiveEffect.
    */
   async rollSave(rollOptions = {}, dialogOptions = {}, messageData = {}, messageOptions = {}) {
     const rollData = this.parent.getRollData();
@@ -157,10 +115,11 @@ export default class BaseEffectModel extends foundry.data.ActiveEffectTypeDataMo
 
     await roll.evaluate();
 
-    if (roll.product) await this.parent.update({ disabled: true });
+    if (roll.product) await this.parent.update({ "duration.expired": true });
 
     messageData.speaker ??= DrawSteelChatMessage.getSpeaker({ actor: this.parent.target });
 
+    messageData.rolls = [roll];
     messageData.type = "standard";
     messageData.system ??= {};
     messageData.system.parts ??= [];